requirements.txt
torch==2.4.1
biopython==1.83
numpy==1.24.4
argparse==1.4.0
